move CLI test class to its own header file
